src/serverless/requirements.txt
azure-functions
tswift